requirements.txt.txt
Django==5.0.2
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
asgiref==3.6.0
dj-database-url==0.5.0
djoser==2.1.0
django-cors-headers==3.13.0
django-debug-toolbar==4.0.0
django-filter==23.1
PyJWT==2.6.0  # Required for JWT authentication